Dev ben

+++ src/lib/net/NetManager.kt
@@ -15,6 +15,7 @@ import java.nio.channels.SelectionKey
 import java.nio.channels.ServerSocketChannel
 import java.nio.channels.SocketChannel
 import lib.net.rw.ReadWriteableSocketChannel
+import java.nio.ByteBuffer
 
 /**
  * Created by bensoer on 04/03/16.
@@ -42,17 +43,38 @@ class NetManager(val addressMapper: AddressMapper): Thread(){
 
     fun addMapping(addressPair:AddressPair):Boolean{
 
-        val serverSocketChannel = NetLibrary.createServerSocket(addressPair.localPort);
-        if(serverSocketChannel == null){
-            Logger.log("NetManager - ERROR - Failed To Create A Server Socket For New Port");
-            return false;
+        if(addressPair.type.equals("UDP")){
+
+
+            val serverSocketChannel = NetLibrary.createUDPServerSocket(addressPair.localPort);
+            if(serverSocketChannel == null){
+                Logger.log("NetManager - ERROR - Failed To Create A UDP Server Socket For New Port");
+                return false;
+            }else{
+                this.addressMapper.addPortMapping(addressPair);
+                this.select.registerChannel(serverSocketChannel);
+                Logger.log("NetManager - UDP Port Mapping Add Complete");
+                return true;
+            }
+
+        }else if(addressPair.type.equals("TCP")){
+
+            val serverSocketChannel = NetLibrary.createServerSocket(addressPair.localPort);
+            if(serverSocketChannel == null){
+                Logger.log("NetManager - ERROR - Failed To Create A Server Socket For New Port");
+                return false;
+            }else{
+                this.addressMapper.addPortMapping(addressPair);
+                this.select.registerServerChannel(serverSocketChannel);
+                Logger.log("NetManager - TCP Port Mapping Add Complete");
+                return true;
+            }
+
         }else{
-            this.addressMapper.addPortMapping(addressPair);
-            this.select.registerServerChannel(serverSocketChannel);
-            Logger.log("NetManager - Mapping Added Complete");
-            return true;
+            Logger.log("NetManager - Requested Mapping Does Not Use An Accepted Protocol Type. Can't Create Mapping");
+            return false;
         }
-        Logger.log("why am i here");
+
     }
 
     fun removeMapping(addressPair:AddressPair):Boolean{
@@ -62,16 +84,14 @@ class NetManager(val addressMapper: AddressMapper): Thread(){
         while(iterator.hasNext()){
             val key = iterator.next();
 
-            println("Number of Keys: ${keysSet.size}");
-
             if(this.select.getChannelForKey(key) is ServerSocketChannel){
+                Logger.log("NetManager:removeMapping - Found ServerSocketChannel to Remove");
                 val channel = this.select.getChannelForKey(key) as ServerSocketChannel;
                 val listeningPort = channel.socket().localPort
 
-                println("Listening Port: $listeningPort");
-
                 //if this ServerSocket is listening on the same port then we know to delete it
                 if(listeningPort == addressPair.localPort){
+                    Logger.log("NetManager:removeMapping - Found Matching Local Ports");
 
                     //close the channel
                     channel.close();
@@ -81,6 +101,24 @@ class NetManager(val addressMapper: AddressMapper): Thread(){
                     //iterator.remove();
 
                     //remove from database
+                    this.addressMapper.deletePortMapping(addressPair);
+
+                    return true;
+                }
+            }else if(this.select.getChannelForKey(key) is DatagramChannel){
+                Logger.log("NetManager:removeMapping - Found DatagramChannel to Remove");
+                val channel = this.select.getChannelForKey(key) as DatagramChannel;
+                val listeningPort = channel.socket().localPort;
+
+                //if this DatagramSocket is listening on the same port then we know to delete it
+                if(listeningPort == addressPair.localPort){
+                    Logger.log("NetManager:removeMapping - Found Matching Local Ports");
+                    //close the channel
+                    channel.close();
+                    //cancel the key
+                    key.cancel();
+
+                    //remove port mapping from database
                     this.addressMapper.deletePortMapping(addressPair);
 
                     return true;
@@ -106,9 +144,18 @@ class NetManager(val addressMapper: AddressMapper): Thread(){
         while(iterator.hasNext()){
             val key = iterator.next();
 
-            //close all sockets
-            val channel = this.addressMapper.getSocketChannel(key);
-            channel?.close();
+            if(key.channel() is SocketChannel){
+                val channel = key.channel() as SocketChannel;
+                channel.close();
+
+            }else if(key.channel() is DatagramChannel){
+                val channel = key.channel() as DatagramChannel;
+                channel.close();
+            }else{
+                Logger.log("NetManager - ERROR - An Invalid Channel Was Detected During Termination. Unable To" +
+                        " Properly Close It.");
+            }
+
 
             //cancel all keys
             key.cancel();
@@ -207,16 +254,11 @@ class NetManager(val addressMapper: AddressMapper): Thread(){
                     val serverSocketSession = serverSocketChannel.accept();
                     if(serverSocketSession == null){
                         Logger.log("NetManager - Nothing To Accept From the Connection")
+                        keyIterator.remove();
                         continue;
                     }
                     Logger.log("NetManager - Connection Accepted");
-                    //serverSocketChannel.configureBlocking(false);
 
-
-                    Logger.log("NetManager - Registering The Channel");
-                    //register this new socket
-                    //val srcKeys = this.select.registerChannel(serverSocketSession);
-                    val srcKeys = this.select.registerChannel(serverSocketSession);
 
                     Logger.log("NetManager - Getting Location To The Forwarding Location From Local Port");
                     //lookup where this is supposed to go
@@ -231,8 +273,8 @@ class NetManager(val addressMapper: AddressMapper): Thread(){
 
                         //close the socket
                         serverSocketSession.close();
-                        //cancel the keys
-                        srcKeys.cancel();
+
+                        keyIterator.remove();
 
                         Logger.log("NetManager - Cleanup Complete. We Are Still Running");
                         continue;
@@ -240,8 +282,12 @@ class NetManager(val addressMapper: AddressMapper): Thread(){
                     //else the complete connection is successful. Register everything for flows
                     }else{
 
+                        Logger.log("NetManager - Registering The Channel");
+                        //register this new socket
+                        val srcKeys = this.select.registerChannel(serverSocketSession);
+
                         //register socket with select
-                        val destKeys = this.select.registerChannel(clientSocketChannel!!);
+                        val destKeys = this.select.registerChannel(clientSocketChannel);
 
                         Logger.log("NetManager - Generating Hashing Maps To Each Socket");
                         //hash the keys to eachothers sockets
@@ -284,6 +330,8 @@ class NetManager(val addressMapper: AddressMapper): Thread(){
                                 addressMapper.deleteKeyForChannel(dataSourceChannel);
                                 key.cancel();
 
+                                keyIterator.remove();
+
                                 //tell statistics of closed socket
                                 this.listener?.connectionClosed();
                                 continue;
@@ -299,11 +347,17 @@ class NetManager(val addressMapper: AddressMapper): Thread(){
                         Logger.log("NetManager - Data Is From A DatagramChannel");
 
                         val dataSourceChannel = key.channel() as DatagramChannel;
-                        val dataSourceRemoteAddress = dataSourceChannel.remoteAddress;
+                        val rwDataSourceChannel = ReadWriteableDatagramChannel(dataSourceChannel);
+
+                        val socketRead = NetLibrary.readFromSocket(rwDataSourceChannel, ByteBuffer.allocate(2048));
+
+                        val dataSourceRemoteAddress = socketRead.sourceAddress;
 
                         //we gotta check whether a mapping exists in a couple places
-                        var dataDestChannel = addressMapper.getDatagramChannel(dataSourceRemoteAddress);
+                        //is this mapping going from C -> PF -> S ?
+                        var dataDestChannel:DatagramChannel? = addressMapper.getDatagramChannel(dataSourceRemoteAddress!!);
                         if(dataDestChannel == null){
+                            //or is this a mapping going from S -> PF -> C ?
                             dataDestChannel = addressMapper.getUDPMapping(dataSourceChannel);
                         }
 
@@ -311,10 +365,14 @@ class NetManager(val addressMapper: AddressMapper): Thread(){
                         if(dataDestChannel == null){
 
                             //means the mapping doesn't exist. assumed a new connection ?
+
+                            //connect it back to the host so we can auto respond properly
+                            dataSourceChannel.connect(dataSourceRemoteAddress);
+
                             val localPort = dataSourceChannel.socket().localPort;
                             val addressPair = this.addressMapper.getPortMapping(localPort, "UDP");
 
-                            val dataDestChannel = NetLibrary.createUDPClientSocket(addressPair!!.dest);
+                            dataDestChannel = NetLibrary.createUDPClientSocket(addressPair!!.dest);
 
                             //if we fail to connect to the forwarding location, closeup shop for this channel
                             if(dataDestChannel == null){
@@ -324,6 +382,8 @@ class NetManager(val addressMapper: AddressMapper): Thread(){
                                 dataSourceChannel.close();
                                 //cancel the source key
                                 key.cancel();
+
+                                keyIterator.remove();
 
                                 Logger.log("NetManager - Cleanup Complete. We Are Still Running");
                                 continue;
@@ -341,13 +401,38 @@ class NetManager(val addressMapper: AddressMapper): Thread(){
                                 //this.addressMapper.createDatagramMapping(dataDestLocalAddress, dataSourceChannel);
 
                                 this.listener?.connectionOpened();
+
+                                //because this is a new UDP channel, its stateless and thus this packet has data to
+                                //be transfered aswell. forward this packet along
+                                NetLibrary.writeToSocket(ReadWriteableDatagramChannel(dataDestChannel), socketRead.data);
+                                val bytesTransfered = socketRead.bytesRead;
+                                listener?.bytesTransfered(dataSourceChannel.remoteAddress as InetSocketAddress, dataSourceChannel.socket().localPort, bytesTransfered);
+                                if(bytesTransfered == -1){
+                                    Logger.log("NetManager - Negative UDP Bytes Transferred. Connection Assumed Closed. Cleaning Up");
+                                    //transfer is done, disconnect
+
+                                    //close the source channel
+                                    this.addressMapper.deleteUDPMappings(dataSourceChannel);
+                                    dataSourceChannel.close();
+                                    key.cancel();
+
+                                    //close the destination channel
+                                    this.addressMapper.deleteUDPMappings(dataDestChannel);
+                                    dataDestChannel.close();
+
+                                    this.listener?.connectionClosed();
+                                }
+
                             }
 
                         }else{
 
                             //means the mapping is known and this is a known connection
 
-                            val bytesTransfered = NetLibrary.transferDataFromChannels(ReadWriteableDatagramChannel(dataSourceChannel), ReadWriteableDatagramChannel(dataDestChannel));
+                            //val bytesTransfered = NetLibrary.transferDataFromChannels(ReadWriteableDatagramChannel(dataSourceChannel), ReadWriteableDatagramChannel(dataDestChannel));
+
+                            NetLibrary.writeToSocket(ReadWriteableDatagramChannel(dataDestChannel), socketRead.data);
+                            val bytesTransfered = socketRead.bytesRead;
                             listener?.bytesTransfered(dataSourceChannel.remoteAddress as InetSocketAddress, dataSourceChannel.socket().localPort, bytesTransfered);
                             if(bytesTransfered == -1){
                                 Logger.log("NetManager - Negative UDP Bytes Transferred. Connection Assumed Closed. Cleaning Up");
@@ -363,29 +448,18 @@ class NetManager(val addressMapper: AddressMapper): Thread(){
                                 dataDestChannel.close();
 
                                 this.listener?.connectionClosed();
+
                             }
-
                         }
-
-
-
 
                     }else{
                         throw IllegalArgumentException("NetManager - No Matching Channel Type For Key's Channel");
                     }
-
-
-
                 }
 
                 keyIterator.remove();
 
             }
-
-
         }
     }
-
-
-
 }
